requirements.txt
BeautifulSoup==3.2.1
Django==1.7
EasyProcess==0.1.6
MySQL-python==1.2.5
Pillow==2.6.0
PyVirtualDisplay==0.1.5
antigate==1.3
argparse==1.2.1
fake-factory==0.4.2
grab==0.4.13
ipython==2.3.0
lxml==3.4.0
pycurl==7.19.5
selenium==2.43.0
wsgiref==0.1.2
xlrd==0.9.3
xlutils==1.7.1
xlwt==0.7.5
xmltodict==0.9.0
